Update rewriting of objective functions
The handling of the objective function as a string has been updated to allow more robust handling of + and - signs as well as whitespaces.

Specifically now supports the usage of variable names that may include whitespaces, eg. "datapoint[i, j]". These are easily encountered when working with pandas input data.

Signed-off-by: Henri Karhula <100681258+henrikarhula@users.noreply.github.com>